- Fixed dataclass from pydantic not working (#100 comment). - Added support for pydantic models and attr defines similar to dataclasses. - Support for python 3.6 will be removed in v5.0.0. New features added in future v4 releases are not guaranteed to work with python 3.6.
