+++ pyproject.toml
@@ -85,6 +85,8 @@ test = [
     "jsonargparse[types-pyyaml]",
     "responses>=0.12.0",
     "types-requests>=2.28.9",
+    "pydantic>=1.10.7; python_version > '3.6'",
+    "attrs>=22.2.0",
 ]
 test-no-urls = [
     "coverage>=4.5.1",
